[FIX] website_sale: we don't support rounding uom in ecommerce
We only support int, following #97a13a004f2cce01294fa22c1157f362a3f654be

opw-2456796

closes odoo/odoo#67286

X-original-commit: a05a34366f6dc8c70b39c15c3441beaf6bbb72d2

+++ addons/website_sale/models/sale_order.py
@@ -152,12 +152,12 @@ class SaleOrder(models.Model):
 
         try:
             if add_qty:
-                add_qty = float(add_qty)
+                add_qty = int(add_qty)
         except ValueError:
             add_qty = 1
         try:
             if set_qty:
-                set_qty = float(set_qty)
+                set_qty = int(set_qty)
         except ValueError:
             set_qty = 0
         quantity = 0
